+++ pakyow-support/CHANGELOG.md
@@ -1,5 +1,10 @@
 # v1.1.0 (unreleased)
 
+  * `chg` **Pass keyword arguments through pipelines.**
+
+    *Related links:*
+    - [Pull Request #425][pr-425]
+
   * `add` **Introduce `#argument_list?` refinement for `Proc`, `Method`, and `UnboundMethod` introspection.**
 
     *Related links:*
@@ -248,6 +253,7 @@
     *Related links:*
     - [Pull Request #364][pr-364]
 
+[pr-425]: https://github.com/pakyow/pakyow/pull/425
 [pr-424]: https://github.com/pakyow/pakyow/pull/424
 [pr-423]: https://github.com/pakyow/pakyow/pull/423
 [pr-422]: https://github.com/pakyow/pakyow/pull/422